requirements.txt
flask
flask-cors
joblib
scikit-learn
pillow
gunicorn